+++ src/scripts/TokenNoteHoverHUD.js
@@ -22,6 +22,8 @@ import { gurps } from "./systems/gurps.js";
 import { wrathAndGlory } from "./systems/wrathAndGlory.js";
 import { dragonbane } from "./systems/dragonbane.js";
 import { worldofdarkness } from './systems/worldofdarkness.js';
+import { forbiddenlands } from './systems/forbiddenlands.js';
+import { deltagreen } from './systems/deltagreen.js';
 
 /**
  * A HUD extension that shows the Note preview
@@ -138,6 +140,10 @@ export default class TokenNoteHoverHUD extends foundry.applications.hud.BasePlac
       tempContent = await dragonbane(actor, displayImages);
     } else if (game.data.system.id === 'worldofdarkness') {
       tempContent = await worldofdarkness(actor, displayImages);
+    } else if (game.data.system.id === 'forbidden-lands') {
+      tempContent = await forbiddenlands(actor, displayImages);
+    } else if (game.data.system.id === 'deltagreen') {
+      tempContent = await deltagreen(actor, displayImages);
     }
 
     this.contentAvailable = tempContent !== null && tempContent !== '';