+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Added support for [Warhammer 40,000 Roleplay: Wrath & Glory](https://foundryvtt.com/packages/wrath-and-glory).
 * Added support for [Dragonbane - Drakar och Demoner](https://foundryvtt.com/packages/dragonbane).
 * Added support for [World of Darkness 20th edition](https://foundryvtt.com/packages/worldofdarkness).
+* Added support for [Forbidden Lands](https://foundryvtt.com/packages/forbidden-lands).
+* Added support for [Delta Green](https://foundryvtt.com/packages/deltagreen).
 
 ## [3.0.17](https://github.com/jendave/token-note-hover/commits/main) (2025-08-14)